+++ packages/sdks/output/solid/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Builder.io Solid SDK Changelog (@builder.io/sdk-solid)
 
+## 3.0.6
+
+### Patch Changes
+
+- abe5cba: Feat: exports `setClientUserAttributes` helper that can be used to set and update Builder's user attributes cookie. This cookie is used by Builder's Personalization Containers to decide which variant to render.
+
+  Usage example:
+
+  ```ts
+  import { setClientUserAttributes } from "@builder.io/sdk-solid";
+
+  setClientUserAttributes({
+    device: "tablet",
+  });
+  ```
+
+## 3.0.5
+
+### Patch Changes
+
+- 91a7117: Fix: vertically aligning child block of columns block
+- 2f73837: Fix: Removed z-index from Video block, which caused it to hide its children elements.
+
 ## 3.0.3
 
 ### Patch Changes